ensure bios never loads OP-TEE pagestore in SDP memory
Move OP-TEE page store load address to 4MB above start of reserved DRAM.
The offset 4MB is related to the first 1MB being reserved for early
boot stage (should be enough) and 2MB reserved to the SRAM emulation
in the memory layout. Offset is 4MB instead of 3MB to make it safer
and aligned on LPAE pgdir granule.

Prior this change, bios loaded the Op-tee pagestore near the end
of the secure DRAM which used as secure data path memory when SDP
support is enabled in OP-TEE.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Tested-by: Jens Wiklander <jens.wiklander@linaro.org>